chore: possibility to not remove logs on remove

+++ logging/subs/remove_performance.sh
@@ -1,6 +1,9 @@
 #! /bin/bash
 # Remove log folder
-sudo rm -rf /var/log/performance/
+read -p "Do you want to remove the logs? (y/n): " is_remove
+if [[ $is_remove == null || $is_remove != "y" ]]; then
+    sudo rm -rf /var/log/performance/
+fi
 
 # Remove script
 sudo rm /usr/local/sbin/performance_logging.sh
